autenticacao-jwt/requirements.txt
ruff
basedpyright
Flask
PyJWT
python-dotenv
types-Flask
mysql-connector-python
bcrypt
pydantic_settings